ai/requirements.txt
torch>=2.0.0
torchvision
facenet-pytorch
opencv-python
numpy
scikit-learn
matplotlib
umap-learn
ultralytics>=8.0.20
requests
